Update generated object files built with new options.
This includes the EOT-for-termination TB trick and the fact that
the code has been rebuilt using a new makefile with new options.

Note these files are included for the convenience of users that
want to try the core without installing an 8051 toolchain. They can
be rebuilt using the supplied makefiles.